requirements.txt
appiumbase~=0.1.0
requests==2.26.0
pytest~=6.2.5
allure-pytest~=2.9.45

